% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Area_Level_Data.R
\docType{data}
\name{Area_Level_Data}
\alias{Area_Level_Data}
\title{Area level data}
\format{
A data frame with 96 rows and 5 columns:
\describe{
\item{RHDA}{Region name}
\item{Percentage_of_immigrants}{percentage of immigrants in each region}
\item{Percentage_of_indigenous}{percentage of indigenous people in each region}
\item{Proporton_of_Low_education}{proportion of persons 15+ who have not graduated high school}
\item{Income}{median household income}
...
}
}
\usage{
Area_Level_Data
}
\description{
The data which describes the sociodemographic characters (proportion of indigenous people, proportions of immigrants, proportion of low education, median household income) for 96 regions.
}
\keyword{datasets}
