% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoWeek.R
\docType{data}
\name{TwoWeek}
\alias{TwoWeek}
\title{TwoWeek}
\format{
A data frame with 700 rows and 2 columns:
\describe{
\item{date_diagnosed}{The date which the disease diagnosed}
\item{V2}{the week}

...
}
}
\usage{
TwoWeek
}
\description{
The simulated data for the date diagnosed and tau
}
\keyword{datasets}
